Add error buffer for LP subprocesses
Otherwise, any error output is hidden from the user, which makes debugging
difficult.

I didn't use the a2ps print menu yet, but its process call is

``` emacs-lisp

(make-process
 :name "a2ps-print"
 :buffer buffer
 :stderr buffer
 :connection-type 'pipe
 :command cmd)
```

so it also shows you the stderr in a buffer. For lp the normal stdout might not
be so useful, so it's fine not having and stdout buffer there.